requirements.txt
numpy==1.24.4
onnx==1.13.1
onnxruntime==1.19.2
Pillow==11.3.0
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.53.1
